requirements.txt
matplotlib==3.7.3
numpy==1.24.1
opencv_python==4.8.1.78
Pillow==9.3.0
Pillow==10.2.0
timm==0.9.7
torch==2.1.0+cu121
tqdm==4.66.1
python==3.8.17
